+++ _goldens/test/_files/event_tearoff.dart
@@ -1,0 +1,13 @@
+import 'package:angular/angular.dart';
+
+@Component(
+  selector: 'uses-event-tearoff',
+  template: r'''
+    <button (click)="onClick" (blur)="onBlur"></button>
+  ''',
+)
+class UsesEventTearoff {
+  void onClick(Object e) {}
+
+  void onBlur() {}
+}